query builder: move post-filter and result-set-modifiers inside projection
