+++ packages/legend-application-query/src/stores/QueryBuilderState.ts
@@ -26,7 +26,6 @@ import {
 } from '@finos/legend-shared';
 import { QueryBuilderFilterState } from './filter/QueryBuilderFilterState.js';
 import { QueryBuilderFetchStructureState } from './fetch-structure/QueryBuilderFetchStructureState.js';
-import { QueryResultSetModifierState } from './fetch-structure/projection/QueryResultSetModifierState.js';
 import {
   QueryTextEditorMode,
   QueryTextEditorState,
@@ -37,7 +36,7 @@ import { QueryBuilderResultState } from './QueryBuilderResultState.js';
 import {
   processQueryBuilderLambdaFunction,
   processQueryParameters,
-} from './QueryBuilderLambdaProcessor.js';
+} from './QueryBuilderValueSpecificationProcessor.js';
 import { QueryBuilderUnsupportedState } from './QueryBuilderUnsupportedState.js';
 import {
   type Class,
@@ -59,20 +58,16 @@ import {
   buildLambdaVariableExpressions,
   buildRawLambdaFromLambdaFunction,
 } from '@finos/legend-graph';
-import { buildLambdaFunction } from './QueryBuilderLambdaBuilder.js';
+import { buildLambdaFunction } from './QueryBuilderValueSpecificationBuilder.js';
 import {
   LambdaParameterState,
   type GenericLegendApplicationStore,
 } from '@finos/legend-application';
-import { QueryBuilderPostFilterState } from './fetch-structure/projection/post-filter/QueryBuilderPostFilterState.js';
-import type { QueryBuilderPostFilterOperator } from './fetch-structure/projection/post-filter/QueryBuilderPostFilterOperator.js';
 import { QueryFunctionsExplorerState } from './explorer/QueryFunctionsExplorerState.js';
 import { QueryParametersState } from './QueryParametersState.js';
-import type { QueryBuilderAggregateOperator } from './fetch-structure/projection/aggregation/QueryBuilderAggregateOperator.js';
 import type { QueryBuilderFilterOperator } from './filter/QueryBuilderFilterOperator.js';
 import { getQueryBuilderCoreFilterOperators } from './filter/QueryBuilderFilterOperatorLoader.js';
-import { getQueryBuilderCorePostFilterOperators } from './fetch-structure/projection/post-filter/QueryBuilderPostFilterOperatorLoader.js';
-import { getQueryBuilderCoreAggregrationOperators } from './fetch-structure/projection/aggregation/QueryBuilderAggregateOperatorLoader.js';
+import { QueryResultSetModifierState } from './fetch-structure/projection/QueryResultSetModifierState.js';
 
 export abstract class QueryBuilderMode {
   abstract get isParametersDisabled(): boolean;
@@ -127,24 +122,18 @@ export class QueryBuilderState {
   queryFunctionsExplorerState: QueryFunctionsExplorerState;
   fetchStructureState: QueryBuilderFetchStructureState;
   filterState: QueryBuilderFilterState;
-  postFilterState: QueryBuilderPostFilterState;
-  resultSetModifierState: QueryResultSetModifierState;
   resultState: QueryBuilderResultState;
   queryTextEditorState: QueryTextEditorState;
   queryUnsupportedState: QueryBuilderUnsupportedState;
   observableContext: ObserverContext;
-  filterOperators: QueryBuilderFilterOperator[] =
-    getQueryBuilderCoreFilterOperators();
-  postFilterOperators: QueryBuilderPostFilterOperator[] =
-    getQueryBuilderCorePostFilterOperators();
-  aggregationOperators: QueryBuilderAggregateOperator[] =
-    getQueryBuilderCoreAggregrationOperators();
+  changeDetectionState: QueryBuilderChangeDetectionState;
   isCompiling = false;
   backdrop = false;
   showFunctionPanel = false;
   showParameterPanel = false;
-  showPostFilterPanel = false;
-  changeDetectionState: QueryBuilderChangeDetectionState;
+
+  filterOperators: QueryBuilderFilterOperator[] =
+    getQueryBuilderCoreFilterOperators();
 
   constructor(
     applicationStore: GenericLegendApplicationStore,
@@ -158,8 +147,6 @@ export class QueryBuilderState {
       queryFunctionsExplorerState: observable,
       fetchStructureState: observable,
       filterState: observable,
-      postFilterState: observable,
-      resultSetModifierState: observable,
       resultState: observable,
       queryTextEditorState: observable,
       queryUnsupportedState: observable,
@@ -168,7 +155,6 @@ export class QueryBuilderState {
       mode: observable,
       showFunctionPanel: observable,
       showParameterPanel: observable,
-      showPostFilterPanel: observable,
       changeDetectionState: observable,
       setMode: action,
       resetQueryBuilder: action,
@@ -178,7 +164,6 @@ export class QueryBuilderState {
       setBackdrop: action,
       setShowFunctionPanel: action,
       setShowParameterPanel: action,
-      setShowPostFilterPanel: action,
       changeClass: action,
       changeFetchStructure: action,
       compileQuery: flow,
@@ -193,11 +178,6 @@ export class QueryBuilderState {
     this.queryFunctionsExplorerState = new QueryFunctionsExplorerState(this);
     this.fetchStructureState = new QueryBuilderFetchStructureState(this);
     this.filterState = new QueryBuilderFilterState(this, this.filterOperators);
-    this.postFilterState = new QueryBuilderPostFilterState(
-      this,
-      this.postFilterOperators,
-    );
-    this.resultSetModifierState = new QueryResultSetModifierState(this);
     this.resultState = new QueryBuilderResultState(this);
     this.queryTextEditorState = new QueryTextEditorState(this);
     this.queryUnsupportedState = new QueryBuilderUnsupportedState(this);
@@ -223,10 +203,6 @@ export class QueryBuilderState {
 
   setShowParameterPanel(val: boolean): void {
     this.showParameterPanel = val;
-  }
-
-  setShowPostFilterPanel(val: boolean): void {
-    this.showPostFilterPanel = val;
   }
 
   getQuery(options?: { keepSourceInformation: boolean }): RawLambda {
@@ -269,11 +245,6 @@ export class QueryBuilderState {
     );
     this.fetchStructureState = fetchStructureState;
     this.filterState = new QueryBuilderFilterState(this, this.filterOperators);
-    this.postFilterState = new QueryBuilderPostFilterState(
-      this,
-      this.postFilterOperators,
-    );
-    this.resultSetModifierState = new QueryResultSetModifierState(this);
     this.fetchStructureState.graphFetchTreeState.initialize();
   }
 
@@ -480,7 +451,8 @@ export class QueryBuilderState {
 
   // TODO-BEFORE-PR: consider removing this
   changeFetchStructure(): void {
-    this.resultSetModifierState = new QueryResultSetModifierState(this);
+    this.fetchStructureState.projectionState.resultSetModifierState =
+      new QueryResultSetModifierState(this.fetchStructureState.projectionState);
     const treeData = this.fetchStructureState.graphFetchTreeState.treeData;
     if (!treeData) {
       this.fetchStructureState.graphFetchTreeState.initialize();
