+++ packages/legend-application-query/src/stores/filter/operators/QueryBuilderFilterOperatorHelper.ts
@@ -36,7 +36,7 @@ import {
 } from '../QueryBuilderFilterState.js';
 import { QUERY_BUILDER_SUPPORTED_FUNCTIONS } from '../../../QueryBuilder_Const.js';
 import { buildGenericLambdaFunctionInstanceValue } from '../../QueryBuilderValueSpecificationBuilderHelper.js';
-import { buildPropertyExpressionChain } from '../../QueryBuilderLambdaBuilder.js';
+import { buildPropertyExpressionChain } from '../../QueryBuilderValueSpecificationBuilder.js';
 import type { QueryBuilderFilterOperator } from '../QueryBuilderFilterOperator.js';
 
 const getPropertyExpressionChainVariable = (
